Separate manifest and source code for each host
