+++ package.json
@@ -16,17 +16,17 @@
     "build-dev": "webpack --mode development",
     "dev-server": "webpack-dev-server --mode development",
     "sideload": "echo Please specify which app using 'npm run sideload:app'.",
-    "sideload:excel": "office-toolbox sideload -m manifest.xml -a excel",
-    "sideload:powerpoint": "office-toolbox sideload -m manifest.xml -a powerpoint",
-    "sideload:word": "office-toolbox sideload -m manifest.xml -a word",
+    "sideload:excel": "office-toolbox sideload -m manifest.excel.xml -a excel",
+    "sideload:powerpoint": "office-toolbox sideload -m manifest.powerpoint.xml -a powerpoint",
+    "sideload:word": "office-toolbox sideload -m manifest.word.xml -a word",
     "start": "office-addin-debugging start manifest.xml",
     "start:desktop": "office-addin-debugging start manifest.xml desktop",
     "start:web": "office-addin-debugging start manifest.xml web",
     "stop": "office-addin-debugging stop manifest.xml",
     "unload": "echo Please specify which app using 'npm run unload:app'.",
-    "unload:excel": "office-toolbox remove -m manifest.xml -a excel",
-    "unload:powerpoint": "office-toolbox remove -m manifest.xml -a powerpoint",
-    "unload:word": "office-toolbox remove -m manifest.xml -a word",
+    "unload:excel": "office-toolbox remove -m manifest.excel.xml -a excel",
+    "unload:powerpoint": "office-toolbox remove -m manifest.powerpoint.xml -a powerpoint",
+    "unload:word": "office-toolbox remove -m manifest.word.xml -a word",
     "validate": "office-toolbox validate -m manifest.xml",
     "watch": "webpack --mode development --watch"
   },
