+++ src/taskpane/project.ts
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
+ * See LICENSE in the project root for license information.
+ */
+
+Office.onReady(info => {
+  if (info.host === Office.HostType.Project) {
+    document.getElementById("sideload-msg").style.display = "none";
+    document.getElementById("app-body").style.display = "flex";
+    document.getElementById("run").onclick = run;
+  }
+});
+
+export async function run() {
+  /**
+   * Insert your Outlook code here
+   */
+}
